:memo: models 수정 (relationship 관계, is_holiday 삭제, Date -> DateTime으로 수정)
# Conflicts:
#	app/modules/auth/models.py
